+++ lib/isodoc/csand/word_convert.rb
@@ -1,5 +1,5 @@
+require_relative "base_convert"
 require "isodoc"
-require_relative "metadata"
 
 module IsoDoc
   module Csand
@@ -31,46 +31,7 @@ module IsoDoc
         }
       end
 
-      def metadata_init(lang, script, labels)
-        @meta = Metadata.new(lang, script, labels)
-      end
-
-      def annex_name(annex, name, div)
-        div.h1 **{ class: "Annex" } do |t|
-          t << "#{get_anchors[annex['id']][:label]} "
-          t.br
-          t.b do |b|
-            name&.children&.each { |c2| parse(c2, b) }
-          end
-        end
-      end
-
-      def term_defs_boilerplate(div, source, term, preface)
-        if source.empty? && term.nil?
-          div << @no_terms_boilerplate
-        else
-          div << term_defs_boilerplate_cont(source, term)
-        end
-      end
-
-      def i18n_init(lang, script)
-        super
-        @annex_lbl = "Appendix"
-      end
-
-      def cleanup(docxml)
-        super
-        term_cleanup(docxml)
-      end
-
-      def term_cleanup(docxml)
-        docxml.xpath("//p[@class = 'Terms']").each do |d|
-          h2 = d.at("./preceding-sibling::*[@class = 'TermNum'][1]")
-          h2.add_child("&nbsp;")
-          h2.add_child(d.remove)
-        end
-        docxml
-      end
+      include BaseConvert
     end
   end
 end
